+++ supabase/functions/resend/index.ts
@@ -2,7 +2,7 @@ import z from 'npm:zod';
 import { corsHeaders } from '../_shared/cors.ts';
 
 const RESEND_API_KEY = Deno.env.get('RESEND_API_KEY');
-const isLocal = Deno.env.get('PUBLIC_SUPABASE_URL')?.includes('host.docker') ?? false;
+const isLocal = Deno.env.get('PUBLIC_SUPABASE_URL')?.includes('127.0.0.1') ?? false;
 
 const ResendBodySchema = z.object({
 	to: z.string().email(),
@@ -46,7 +46,7 @@ const handler = async (request: Request): Promise<Response> => {
 					from: 'notifications@reciprocal.reviews',
 					to: to,
 					subject: subject,
-					html: message
+					text: message
 				})
 			});